requirements.txt
awscli==1.22.34
bcc==0.18.0
beautifulsoup4==4.12.3
blinker==1.7.0
botocore==1.23.34
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
docutils==0.17.1
Flask==3.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==0.10.0
kazam==1.4.5
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.1.4
more-itertools==8.10.0
natsort==8.0.2
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
Pillow==9.0.1
pop-transition==1.1.2
powerline-gitstatus==1.3.1
powerline-status==2.7
protobuf==3.12.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.5
StrEnum==0.4.15
systemd-python==234
terminator==2.1.3
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
